% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{background}
\alias{background}
\title{Background data}
\format{A dataset with 470 rows and 22 columns.
\describe{
  \item{Subject}{ID for each pupil.}
  \item{Class}{ID for each class.}
  \item{LanguageGroup}{Whether the child belonged to the Portuguese-speaking control group, the French-speaking control group, the German-speaking control group, the French-Portuguese bilingual group or the German-Portuguese bilingual group.}
  \item{HLC}{Did the child take heritage language and culture classes?}
  \item{AgeAtT1}{Age of the child at the first data collection.}
  \item{NumberOfLanguages}{Number of languages spoken by the child.}
  \item{Language1, Language2}{The first and second native language spoken by the child. The difference between Swiss-German and Standard German wasn't maintained systematically. (P = Portuguese, F = French, D = German, CH-D = Swiss-German, E = English, I = Italian)}
  \item{LanguageOther1, LanguageOther2}{Other languages known by the child.}
}}
\usage{
background
}
\description{
Background information about the pupils.
}
\references{
Lambelet, Amelia, Raphael Berthele, Magalie Desgrippes, Carlos Pestana and
Jan Vanhove. 2018. Testing interdependence in Portuguese heritage speakers in
Switzerland: The HELASCOT project. In Raphael Berthele and Amelia Lambelet
(Eds.), Heritage language and school language literacy development in migrant
children: Interdependence or independence? 26-33. Bristol, UK: Multilingual
Matters. URL http://doi.org/10.21832/BERTHE9047
}
\keyword{datasets}
