% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{helascot}
\alias{helascot}
\alias{helascot-package}
\title{HELASCOT datasets}
\description{
Datasets for the Heritage Language and School Language (HELASCOT) project.
}
\references{
Berthele, Raphael and Amelia Lambelet (Eds.). 2018. \emph{Heritage and school language literacy development in migrant children: Interdependence or independence?} Bristol, UK: Multilingual Matters. \url{http://doi.org/10.21832/BERTHE9047}.

Vanhove, Jan, Audrey Bonvin, Amelia Lambelet and Raphael Berthele. 2019. Predicting perceptions of the lexical richness of short French, German, and Portuguese texts. \emph{Journal of Writing Research} 10(3). 499-525. \url{http://doi.org/10.17239/jowr-2019.10.03.04}.
}
