% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ratings}
\alias{ratings}
\title{Lexical richness ratings}
\format{A dataset with 29179 rows and 19 columns.}
\source{
\url{https://osf.io/6bywg/}
}
\usage{
ratings
}
\description{
Individual ratings of lexical richness on a 1-9 scale
for 1000 French, 1020 German and 1040 Portuguese texts
by 574 raters. The raters rated around 50 texts each.
}
\references{
Vanhove, Jan, Audrey Bonvin, Amelia Lambelet and Raphael Berthele. 2019.
Predicting perceptions of the lexical richness of short French, German, and
Portuguese texts. Journal of Writing Research 10(3). 58-82. URL
http://doi.org/10.17239/jowr-2019.10.03.04
}
\keyword{datasets}
