% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skills}
\alias{skills}
\title{Reading and writing skill data}
\format{A dataset with 1904 rows and 6 columns.}
\usage{
skills
}
\description{
The pupils' performance on the reading and writing tasks.
}
\references{
Pestana, Carlos, Amelia Lambelet and Jan Vanhove. 2018. Reading comprehension
development in Portuguese heritage speakers in Switzerland (HELASCOT project).
In Raphael Berthele and Amelia Lambelet (Eds.), Heritage language and school
language literacy development in migrant children: Interdependence or
independence? 58-82. Bristol, UK: Multilingual Matters.
URL http://doi.org/10.21832/BERTHE9047

Desgrippes, Magalie, Amelia Lambelet and Jan Vanhove. 2018. The development of
argumentatitive and narrative writing skills in Portuguese heritage speakers in
Switzerland (HELASCOT project). In Raphael Berthele and Amelia Lambelet (Eds.),
Heritage language and school language literacy development in migrant children:
Interdependence or independence? 83-96. Bristol, UK: Multilingual Matters.
URL http://doi.org/10.21832/BERTHE9047
}
\keyword{datasets}
