% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{text_metrics}
\alias{text_metrics}
\title{Text-based metrics}
\format{A dataset with 3060 rows and 137 columns.}
\source{
\url{https://osf.io/6bywg/}
}
\usage{
text_metrics
}
\description{
Metrics of lexical diversity, lexical sophistication,
and related constructs for each of the 3,060 texts rated
for their lexical richness. See \code{\link{ratings}} for the ratings.
}
\references{
Vanhove, Jan, Audrey Bonvin, Amelia Lambelet and Raphael Berthele. 2019.
Predicting perceptions of the lexical richness of short French, German, and
Portuguese texts. Journal of Writing Research 10(3). 58-82. URL
http://doi.org/10.17239/jowr-2019.10.03.04
}
\keyword{datasets}
